Map fragment added (Hope it works)

+++ code/app/src/main/java/com/example/project007/Location.java
@@ -1,15 +1,17 @@
 package com.example.project007;
 
 public class Location {
-    float longitude;
-    float latitude;
+    double longitude;
+    double latitude;
 
-    public Location(float longitude, float latitude) {
+    public Location(double longitude, double latitude) {
         this.longitude = longitude;
         this.latitude = latitude;
     }
 
-    public float getLongitude() {
+
+
+    public double getLongitude() {
         return longitude;
     }
 
@@ -17,7 +19,7 @@ public class Location {
         this.longitude = longitude;
     }
 
-    public float getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
